[Driver] Remove unused input type code (NFCI)
Summary:
When the Swift driver executable was first added in
https://github.com/apple/swift/commit/ed2038585f0c25bb2dd1cec35da565df2df48c82,
the code included a FIXME to support input types, like Clang's
`-x <language>` option. Perhaps as a first step in implementing this
functionality, it also included dead code that manipulated an input type
argument, even though the variable for storing the argument was never
written to.

Remove the unused input type code, and the FIXME. The FIXME is now
tracked with a Swift bug URL, so that a discussion on next steps can be had:
https://bugs.swift.org/browse/SR-6054

Test plan: `utils/build-script --release --test`