RemoteAST: only swift has nominal type descriptors
When the RemoteAST is used to access objects from LLDB, the object that
we are interacting with may not be a Swift object but rather an ObjC
object that has been bridged into swift.  In such a case, the object
does not have a nominal type descriptor associated with it.  Ensure that
the type is a valid Swift type before reading the nominal type
descriptor from the metadata.  This avoids an assertion when evaluating
expressions in LLDB.