Add a new option -sil-serialize-early
This option is used to serialize SIL right before the performance inlining. The aim is to serialize a "high-level" SIL, which still contains all the calls of `@_semantics` functions. This preserves more information and allows for better optimizations of the clients.

Currently only the standard library is supposed to be built with this option.