Possible fix for SR-1226.
Allow calls to explicitly unavailable code when the caller has the
same unavailability.

For example, in a framework compiled with `-application-extension`, this should
be allowed:

	@available(iOSApplicationExtension, unavailable)
	    func x() {
            _ = UIApplication.shared
	    }
	}

since the consumer of the framework, or other call sites within the same
framework will be unable to call the unsafe wrapper in the same situations they
couldn't call the original unsafe code.