[AST] Drop unnecessary friend declaration
Summary:
The friend declaration wasn't qualified (it should have included the
namespace, i.e. `syntax::LegacyASTTransformer`). The effect was instead
to declare a `LegacyASTTransformer` inside the `swift` namespace, which
would then trip up Windows compilation because of non-conforming
two-phase lookup. The implication is that the friend declaration is
actually unneeded, since the incorrect declaration wasn't causing any
other problems, so we can just drop it.