IRGen: do not use the Large Code Model on Win64
The large code model is not necessary.  Large Address Aware should be
sufficient for dealing with the higher address ranges.  Using the large
code model breaks the DLL storage semantics on Windows.  With this we
are able to build most of the standard library for all variants of
Windows.

Addresses SR-6150!