stdlib: add support for AAPCS64 variadics
The AAPCS64 ABI requires that `va_list` be defined as:

    typedef struct __va_list {
      void *__stack;  /* next stack param */
      void *__gr_top; /* end of GP reg save area */
      void *__vr_top; /* end of FP/SIMD reg save area */
      int __gr_offs;  /* offset from __gr_top to next GP register arg */
      int __vr_offs;  /* offset from __vr_top to next FP register arg */
    } va_list;

This type does not decay and does not match pointer size.

`__stack` points to the register save area.  `__gr_top` points to the
byte immediately following the GPR save area, which is rounded to a
16-byte boundary.  `__vr_top` points to the byte immediately following
the FPR save area, which is rounded to a 16-byte boundary.  `__gr_offs`
contains the `((8 - GRCount) * 8)`, `__vr_offs` contains
`((8 - FPCount) * 16)`.