[dead-function-elimination] Do not consider transparent functions as anchors
Transparent functions are emitted into the client in most cases, so there is no need to keep them alive in the defining module unless they are used there.

The only exception are transparent global initializers, because bodies of global initializer functions are never SIL serialized and are always IRGened into the defining module, so that clients can invoke them

This change results in small code size reductions of the binaries being produced after recent changes related to the early SIL serialization and the removal of -sil-serialize-all mode.

Fixes rdar://problem/35022861