[pred-memopts] Update predictable mem opts for ownership.
We support running in both modes for now. The only differences in the two modes
are that:

1. That the ownership version hoists the copies from the load [copy]
early. The non-ownership version just propagates the value and relies on the
retain that is left in place (since without ownership load [copy] is split in 2
instructions).

2. The ownership version (as of this commit), does not support optimizing load
[take]. This is because the manner in which the non-ownership version optimizes
load [take] is **100% unsafe** and thus triggers the ownership verifier,
YAY!. We have just been lucky not to hit any miscompiles from this in
practice. To fix this, I will need to extend the pass in a way that is more than
"updating the pass for ownership" so I am going to add support in a subsequent
commit. This is significant enough work that it really needs to be scheduled
appropriately, so I am not going to finish implementing it right now (I have a
partial implementation). I do not expect any issues from splitting this work in
such a manner since all of the test cases in tree for predictable memopts are
load [copy] test cases.

For more information on 2, see below.

rdar://31521023

----

In the non-ownership verification case, we (as you can imagine) do not care
about double consumes or ownership correctness. Instead, we rely on
strong_retains being in the appropriate places to preserve lifetimes and /hope/
that the frontend/optimizer does not do anything too crazy to cause us to
miscompile. This cavalier attitude results in us forwarding available values
while leaving the store in place causing a region of code to exist where we have
only incrementing the reference count once of the value, but have two copies of
the value: the SSA value and the stored value, i.e. 1 lifetime for two values
=><=. With ownership this can not be done since the ownership verifier
(properly) identifies the double consume and asserts. Lets look at an
example. Consider the following:

     class Klass {}
     sil @foo : $@convention(thin) (@owned Klass) -> () {
     bb0(%0 : @owned $Klass):
       %1 = alloc_stack $Klass
       store %0 to [init] %1 : $*Klass
       ...
       cond_br %_, bb1, bb2

     bb1:
       ...
       %2 = load [take] %1 : $*Klass
       ...
       apply %use_klass(%2) : $@convention(thin) (@owned Klass) -> ()
       ...

     bb2:
       ...
       apply %in_klass_use(%1) : $@convention(thin) (@in Klass) -> ()
       ...
     }

Without ownership, we would transform this to:

     sil @foo : $@convention(thin) (@owned Klass) -> () {
     bb0(%0 : @owned $Klass):
       %1 = alloc_stack $Klass
       store %0 to [init] %1 : $*Klass
       ...
       cond_br %_, bb1, bb2

     bb1:
       ...
       apply %use_klass(%0) : $@convention(thin) (@owned Klass) -> ()
       ...

     bb2:
       ...
       apply %in_klass_use(%1) : $@convention(thin) (@in Klass) -> ()
       ...
     }

Notice that we have not eliminated the store to %1 implying that in BB1, we have
two values that are guarded by the same reference count increment and thus have
the same "lifetime". This is fundamentally unsafe and the ownership verifier
properly flags this as a double consume operation. The only way that we can
safely perform said operation is if we can eliminate the first store, naively:

     sil @foo : $@convention(thin) (@owned Klass) -> () {
     bb0(%0 : @owned $Klass):
       %1 = alloc_stack $Klass
       ...
       cond_br %_, bb1, bb2

     bb1:
       ...
       apply %use_klass(%0) : $@convention(thin) (@owned Klass) -> ()
       ...

     bb2:
       ...
       apply %in_klass_use(%1) : $@convention(thin) (@in Klass) -> ()
       ...
     }

This is a miscompile since we have not considered that there may be other write
dependencies on the store to %1 along other paths since bb1 does not
post-dominate the store. So to implement this correctly, we need to place a
store at the beginning of bb2, i.e.:

     sil @foo : $@convention(thin) (@owned Klass) -> () {
     bb0(%0 : @owned $Klass):
       %1 = alloc_stack $Klass
       ...
       cond_br %_, bb1, bb2

     bb1:
       ...
       apply %use_klass(%0) : $@convention(thin) (@owned Klass) -> ()
       ...

     bb2:
       store %0 to [init] %1 : $*Klass
       ...
       apply %in_klass_use(%1) : $@convention(thin) (@in Klass) -> ()
       ...
     }

As an additional condition, we also need to be sure that we are able to properly
eliminate all read dependencies that are along the path from the store to
%use_klass.

So to implement this safely, we need to:

1. Process loads in topological order by doing an RPO traversal over the CFG.
2. Invalidate available values that have been read when trying to eliminate load
[take]. This is a separate bit since in the load [copy] case, we do not care
about this constraint.
3. Modify the loop handling code to not just clear all availability information
when we have a conflict. We have to be sure to propagate along the read bit so
that any load [take] after such loops are not eliminated.
4. When we try to optimize a load [take], walk up the CFG and identify the
blocks that cause post-dominance failure and insert new stores as the first
instruction in those blocks and then remove the original store.

As you can imagine, this is a non-trivial amount of work that given where I am
in my planned work and that all test cases for this pass are load [copy] test
cases (implying it is not /that/ important) that will be scheduled later and
done in a subsequent commit.

In the meantime, this unsafe optimization will be eliminated and we have found
another case where the ownership verifier has helped us to improve the Swift
Compilers quality.